Support newly-added input event system call (#1)
This commit adds the support for the new event system in rv32emu. By
invoking system call poll_event to accept input specific events, we can
now move the character and control the menu with the keyboard, adjust
the view angle with the mouse in relative mode and use the mouse to
control weapon.